refactor: simplify code structure following thermo-nuclear quality review
- Delete getPasswordResetTokenByHash and markPasswordResetTokenUsed from db.ts and routes to reduce data-model leakage
- Redesign resetPasswordWithToken to run self-contained token verification, consumption, and password updates inside a single database transaction
- Centralize SQLite UNIQUE constraint error handling into a generic handleDbError utility function in routes/util.ts
- Refactor routes/users.ts to use the new handleDbError, eliminating duplicated error response mappings
